requirements.txt
click
python-digitalocean
PyYAML
